images/logging_test/requirements.txt
adlfs
prefect-shell
azure-identity
azure-keyvault-secrets
singer-python